--- lib/geld/excise/version.rb
@@ -1,5 +1,0 @@
-module Geld
-  module Excise
-    VERSION = "0.3.1"
-  end
-end
