+++ lib/jct/version.rb
@@ -1,0 +1,3 @@
+module Jct
+  VERSION = "0.3.1"
+end
