+++ test/test_helper.rb
@@ -1,5 +1,5 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 $LOAD_PATH.unshift(__dir__)
 
-require 'geld/excise'
+require 'jct'
 require "minitest/autorun"
